Move contents from export_test to package_test, delete file
